requirements-dev.txt
black==23.3.0
wheel==0.40.0
pre-commit==3.2.2
mypy-boto3-ec2==1.26.106
mypy-boto3-ecs==1.26.116
mypy-boto3-iam==1.26.97
mypy-boto3-identitystore==1.26.107
mypy-boto3-organizations==1.26.83
mypy-boto3-sts==1.26.57
mypy-boto3-servicecatalog==1.26.109
mypy-boto3-sso-admin==1.26.63
